:recycle: refactor: 장바구니는 언제든 신청이 되도록 수정

+++ src/main/java/com/v1/skuproject/config/jwt/JwtProvider.java
@@ -33,7 +33,7 @@ public class JwtProvider {
     /**
        Access Token 생성
      */
-    public String generateToken(Long userId, int studentId) {
+    public String generateToken(Long userId, long studentId) {
         return Jwts.builder()
                 .claim("userId", userId)
                 .claim("studentId", studentId)
